Update application Status Enum to exclude cancelled status from IDIR_ACTIVE_APPLICATION_STATUS
